+++ test/BtCancelMessageTest.cc
@@ -33,14 +33,15 @@ public:
   class MockBtMessageDispatcher2 : public MockBtMessageDispatcher {
   public:
     size_t index;
-    uint32_t begin;
-    size_t length;
+    int32_t begin;
+    int32_t length;
   public:
     MockBtMessageDispatcher2():index(0),
                                begin(0),
                                length(0) {}
 
-    virtual void doCancelSendingPieceAction(size_t index, uint32_t begin, size_t length) {
+    virtual void doCancelSendingPieceAction
+    (size_t index, int32_t begin, int32_t length) {
       this->index = index;
       this->begin = begin;
       this->length = length;
@@ -60,8 +61,8 @@ void BtCancelMessageTest::testCreate() {
   SharedHandle<BtCancelMessage> pm = BtCancelMessage::create(&msg[4], 13);
   CPPUNIT_ASSERT_EQUAL((uint8_t)8, pm->getId());
   CPPUNIT_ASSERT_EQUAL((size_t)12345, pm->getIndex());
-  CPPUNIT_ASSERT_EQUAL((uint32_t)256, pm->getBegin());
-  CPPUNIT_ASSERT_EQUAL((size_t)1024, pm->getLength());
+  CPPUNIT_ASSERT_EQUAL(256, pm->getBegin());
+  CPPUNIT_ASSERT_EQUAL(1024, pm->getLength());
 
   // case: payload size is wrong
   try {
